pyproject.toml
[project]
name = "charwork"
version = "0.1.0"
description = "Hanzi Project migrated to FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.11.0",
    "aiomysql>=0.2.0",
    "aio-pika>=9.4.0",
    "redis>=5.0.0",
    "elasticsearch>=8.11.0,<9.0.0",
    "celery>=5.3.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.6",
    "volcengine>=1.0.0",
    "bcrypt>=4.0.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pypinyin>=0.53.0",
]

[tool.setuptools]
packages = ["app"]
